Import/Export catmaid connection details to system variables (e.g. for tests)

catmaid_connection_setenv sets environment variables based on a catmaid_connection object.

catmaid_connection_getenv fetches appropriately named environment variables and uses them to open a catmaid connection.

catmaid_connection_unsetenv unsets the environment variables.

Fetch connection details from appropriate environment variables

Unset catmaid connection environment variables

catmaid_connection_setenv(conn = NULL, ...)

catmaid_connection_getenv(...)

catmaid_connection_unsetenv()

Arguments

conn

A catmaid_connection object. The default value of NULL implies that the most recent cached open connection will be used.

...

additional arguments passed to catmaid_login

Value

catmaid_connection_setenv returns TRUE or FALSE depending on whether variables were set successfully. catmaid_connection_getenv returns a connection object created based on environment variables.

Details

catmaid_connection_setenv will attempt to login if this has not already been done

See also

catmaid_login