Convert between entity (neuron) and model (skeleton) ids

catmaid_entities_from_models converts from neuron ids to skeleton ids

catmaid_models_from_entities converts skeleton ids to from neuron ids

catmaid_entities_from_models(skids, pid = 1, conn = NULL, ...)

catmaid_models_from_entities(nids, pid = 1, conn = NULL, ...)

Arguments

skids

One or more numeric skeleton ids or a character vector defining a query (see catmaid_skids or examples for the syntax).

pid

Project id (default 1)

conn

A catmaid_connection objection returned by catmaid_login. If NULL (the default) a new connection object will be generated using the values of the catmaid.* package options as described in the help for catmaid_login.

...

Additional arguments passed to the catmaid_fetch function

nids

Neuron (entity) ids

Value

For catmaid_entities_from_models an integer vector of entity ids each named by the corresponding model id (usually a skeleton id). For catmaid_models_from_entities a list of model ids (usually skeleton ids) named by the corresponding entity ids. Formally there may be multiple skeletons per neuron.

Details

This will normally be used to turn skeleton ids into neuron ids which are used e.g. for annotation purposes. This is probably not something that many end users will need but is required e.g. by catmaid_remove_annotations_for_skeletons.

See also

Examples

ents=catmaid_entities_from_models('ORN PNs left', conn=vfbcatmaid('l1em'))
ents
#>  3486381  8021456  4354933  4493199  7843730  8399890  7515491  7092417 
#>  3486382  4137843  4354934  4493200  5478205  6703262  7001631  7092418 
#>  7107433  7198338  7852250  7845346  7851153  7864823  7865652  7865696 
#>  7107434  7198339  7845338  7845347  7851154  7864824  7865653  7865697 
#>  7865767  7873135  7916593  7934904 16259595 
#>  7865768  7873136  7916873  7934905 16259596 
mods=catmaid_models_from_entities(unname(ents), conn=vfbcatmaid('l1em'))
str(mods)
#> List of 21
#>  $ 3486382 : int 3486381
#>  $ 4137843 : int 8021456
#>  $ 4354934 : int 4354933
#>  $ 4493200 : int 4493199
#>  $ 5478205 : int 7843730
#>  $ 6703262 : int 8399890
#>  $ 7001631 : int 7515491
#>  $ 7092418 : int 7092417
#>  $ 7107434 : int 7107433
#>  $ 7198339 : int 7198338
#>  $ 7845338 : int 7852250
#>  $ 7845347 : int 7845346
#>  $ 7851154 : int 7851153
#>  $ 7864824 : int 7864823
#>  $ 7865653 : int 7865652
#>  $ 7865697 : int 7865696
#>  $ 7865768 : int 7865767
#>  $ 7873136 : int 7873135
#>  $ 7916873 : int 7916593
#>  $ 7934905 : int 7934904
#>  $ 16259596: int 16259595