Return information about connectors joining sets of pre/postsynaptic skids

Return information about connectors joining sets of pre/postsynaptic skids

catmaid_get_connectors_between(
  pre_skids = NULL,
  post_skids = NULL,
  get_names = FALSE,
  pid = 1,
  conn = NULL,
  raw = FALSE,
  ...
)

Arguments

pre_skids, post_skids
Skeleton ids in any form understood by catmaid_skids or NULL meaning no restriction.

get_names
Whether to fetch the neuron name for each pre- and post-synaptic skid (default FALSE).

pid
project id (default 1)

conn
the catmaid_connection object

raw
Whether to return completely unprocessed data (when TRUE) or to convert the nodes and connectors lists into processed data.frames (when FALSE, the default)

...
Additional arguments passed to the catmaid_fetch function.

Value

A data.frame with columns

pre_skid

post_skid

connector_id

pre_node_id

post_node_id

connector_x

connector_y

connector_z

pre_node_x

pre_node_y

pre_node_z

post_node_x

post_node_y

post_node_z

pre_confidence

pre_user

post_confidence

post_user

pre_name (optional - the name of the presynaptic neuron)

post_name (optional - the name of the postsynaptic neuron)

Details

If either the pre_skids or post_skids arguments are not specified (taking the default NULL value) then this implies there is no restriction on the pre- (or post-) synaptic partners.

Each row is a unique set of pre_synaptic node, post_synaptic node, connector_id. A rare (and usually erroneous) scenario is if the same pre_node and post_node are present with two different connector_ids - this would create two rows.

See also