Get statistics for all labels in a project from a CATMAID server

catmaid_get_label_stats returns a data.frame with one row for every node-tag pair.

catmaid_get_all_labels returns a character vector of labels

catmaid_get_label_stats(pid = 1, conn = NULL, ...)

catmaid_get_all_labels(pid = 1, conn = NULL, ...)

Arguments

pid
Project id (default 1)

conn
A catmaid_connection objection returned by catmaid_login. If NULL (the default) a new connection object will be generated using the values of the catmaid.* package options as described in the help for catmaid_login.

...
Additional arguments passed to the catmaid_fetch function

See also

Examples

label_stats=catmaid_get_label_stats()
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:nat’:
#> 
#>     intersect, setdiff, union

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# select soma labels
soma_labels=label_stats %>%
  filter(labelName=='soma') %>%
  group_by(skeletonID)

# select skeleton ids for neurons with multiple cell bodies
multiple_soma=soma_labels %>%
  count(skeletonID) %>%
  filter(n>1) %>%
  arrange(desc(n))

multiple_soma_info = soma_labels %>%
  filter(skeletonID%in% multiple_soma$skeletonID)

head(catmaid_get_all_labels())

#> [1] ">:(" ":)"  "!"   "?"   "?!"  "???"