Get/Set/Remove CATMAID treenode labels (aka tags)

catmaid_get_labels gets labels (tags) for specified nodes.

catmaid_set_labels sets labels (tags) for a specified node.

catmaid_remove_labels removes labels (tags) for a specified node.

catmaid_get_labels(treenodes = NULL, connectors = NULL, pid = 1,
  conn = NULL, ...)

catmaid_set_labels(node, labels, type = c("treenode", "connector"),
  delete_existing = FALSE, pid = 1, conn = NULL, ...)

catmaid_remove_labels(node, labels, type = c("treenode", "connector"),
  pid = 1, conn = NULL, ...)

Arguments

treenodes, connectors
Integer ids of nodes to be queried

pid
Project id (default 1)

conn
A catmaid_connection objection returned by catmaid_login. If NULL (the default) a new connection object will be generated using the values of the catmaid.* package options as described in the help for catmaid_login.

...
Additional arguments passed to the catmaid_fetch function

node
A single tree or connector node to modify

labels
A character vector specifying the labels to add or remove

type
A character vector specifying the type of node to modify

delete_existing
For catmaid_set_labels whether to remove all existing labels before setting the new ones.

Value

catmaid_get_label_stats returns a data.frame with columns

id (integer)

type (treenode or connector)

label (the tag)

A zero row data.frame will be returned if there are no results.

catmaid_remove_labels and catmaid_set_labels throw an error on failure or invisibly return a list containing status information.

See also