Get review status of neurons from CATMAID

Get review status of neurons from CATMAID

catmaid_get_review_status(skids, pid = 1, conn = NULL, ...)

Arguments

skids

One or more numeric skeleton ids or a character vector defining a query (see catmaid_skids or examples for the syntax).

pid

Project id (default 1)

conn

A catmaid_connection objection returned by catmaid_login. If NULL (the default) a new connection object will be generated using the values of the catmaid.* package options as described in the help for catmaid_login.

...

Additional arguments passed to the catmaid_fetch function

Value

a data.frame consisting of total number of nodes and number of reviewed nodes. Note that should any neurons be invalid a warning will be generated and a row with NA values will be returned.

See also

catmaid_fetch

Examples

## Not run: ------------------------------------
# catmaid_get_review_status(skids=c(10418394,4453485))
## ---------------------------------------------