Fetch or translate catmaid users for current/specified connection/project

catmaid_userids convert login names to numeric ids

catmaid_get_user_list(pid = 1, conn = NULL, ...)

catmaid_userids(x, pid = 1, conn = NULL, ...)

Arguments

pid

Project id (default 1)

conn

A catmaid_connection objection returned by catmaid_login. If NULL (the default) a new connection object will be generated using the values of the catmaid.* package options as described in the help for catmaid_login.

...

Additional arguments passed to the catmaid_fetch function

x

A character vector of login names. If x contains valid integer (or numeric) ids then they will be returned as is.

Value

For catmaid_get_user_list a data.frame with For catmaid_userids an integer vector of numeric ids

See also

Examples

# \donttest{
head(catmaid_get_user_list())
#> $longname
#> [1] ""
#> 
#> $userid
#> [1] 2
#> 
#> $username
#> [1] "AnonymousUser"
#> 
#> $is_superuser
#> [1] FALSE
#> 
#> $is_authenticated
#> [1] FALSE
#> 
#> $userprofile
#> $userprofile$independent_ontology_workspace_is_default
#> [1] FALSE
#> 
#> $userprofile$show_text_label_tool
#> [1] FALSE
#> 
#> $userprofile$show_tagging_tool
#> [1] FALSE
#> 
#> $userprofile$show_cropping_tool
#> [1] FALSE
#> 
#> $userprofile$show_tracing_tool
#> [1] TRUE
#> 
#> $userprofile$show_ontology_tool
#> [1] FALSE
#> 
#> $userprofile$show_roi_tool
#> [1] FALSE
#> 
#> $userprofile$primary_group_id
#> NULL
#> 
#> 
# }
# \donttest{
catmaid_userids(1)
#> [1] 1
catmaid_userids('greg')
#> NULL
# }