Fetch a data frame containing details of all volumes on CATMAID server

Fetch a data frame containing details of all volumes on CATMAID server

catmaid_get_volumelist(conn = NULL, pid = 1, ...)

Arguments

conn

the catmaid_connection object

pid

project id (default 1)

...

Additional arguments passed to the catmaid_fetch function.

Value

Since CATMAID v2018.07.19-457-ga09910b a data.frame with columns

id

name

comment

user_id

editor_id

project_id

creation_time

edition_time

annotations

See also

Examples

# \donttest{
vl <- catmaid_get_volumelist()
head(vl)
#>   id name                                  comment user_id editor_id project_id
#> 1  2  cns Imported HDF5 mesh (project: 1 stack: 3)     156       156          1
#>              creation_time             edition_time annotations        area
#> 1 2017-01-03T20:10:29.528Z 2017-01-03T20:10:29.528Z          NA 66791800000
#>        volume watertight meta_computed
#> 1 8.27854e+14      FALSE          TRUE
# find meshes whose name matches neuropil
subset(vl, grepl('neuropil', name))
#>  [1] id            name          comment       user_id       editor_id    
#>  [6] project_id    creation_time edition_time  annotations   area         
#> [11] volume        watertight    meta_computed
#> <0 rows> (or 0-length row.names)
# }