Get, query, set and remove CATMAID meta-annotations for annotations

catmaid_get_meta_annotations Query by annotations to find the meta-annotations that label them.

catmaid_query_meta_annotations Query by meta-annotations to find the annotations that they label.

catmaid_set_meta_annotations Meta-annotate a group of CATMAID annotations

catmaid_remove_meta_annotations Remove meta-annotations from annotations.

catmaid_get_meta_annotations(annotations, pid = 1, conn = NULL, ...)

catmaid_query_meta_annotations(
  meta_annotations,
  with_annotations = FALSE,
  pid = 1,
  conn = NULL,
  ...
)

catmaid_set_meta_annotations(
  meta_annotations,
  annotations,
  pid = 1,
  conn = NULL,
  ...
)

catmaid_remove_meta_annotations(
  annotations,
  meta_annotations,
  force = FALSE,
  pid = 1,
  conn = NULL,
  ...
)

Arguments

annotations
annotation ids designating which annotations to meta-annotate. IDs can be found by calling catmaid_get_annotationlist. If a character string is given, then IDs will be found by calling catmaid_get_annotationlist.

pid
Project id (default 1)

conn
A catmaid_connection objection returned by catmaid_login. If NULL (the default) a new connection object will be generated using the values of the catmaid.* package options as described in the help for catmaid_login.

...
Additional arguments passed to the catmaid_fetch function

meta_annotations
meta-annotation to add to query. Either a vector of IDs or a character sting of meta-annotations can be given.

with_annotations
whether or not to return the other meta-annotations of an annotation, when using catmaid_query_meta_annotations.

force
Whether to force the catmaid server to remove multiple annotations (default FALSE, to provide some protection against accidents).

See also

Examples

if (FALSE) {
## Against FAFB CATMAID server
catmaid_query_meta_annotations("ItoLee_Lineage")

# note that this is similar to:
catmaid_query_by_annotation("^ItoLee_Lineage$", type = 'annotation')
}