Parse a CATMAID URL to extract parameters including xyz location

Parse a CATMAID URL to extract parameters including xyz location

catmaid_parse_url(x)

Arguments

x

One or more catmaid urls

Value

A data.frame containing parsed parameters. The columns will be named according to CATMAID's convention except for the x, y and z location (named as such rather than xp, yp, zp as returned by CATMAID).

Examples

u=paste0("https://neuropil.janelia.org/tracing/fafb/v12/?pid=1&zp=177100&",
  "yp=166068&xp=373086&tool=tracingtool&sid0=7&s0=1.000000")
catmaid_parse_url(u)

#>                                         baseurl pid      z      y      x
#> 1 https://neuropil.janelia.org/tracing/fafb/v12   1 177100 166068 373086
#>          tool sid0 s0
#> 1 tracingtool    7  1

## Not run: ------------------------------------
# cpu=catmaid_parse_url(u)
# n=read.neuron.catmaid(cpu$active_skeleton_id)
# plot3d(n)
# spheres3d(xyzmatrix(cpu), r=1000, col='magenta')
## ---------------------------------------------