Rename (presently) a single neuron or other entity (e.g. annotation)

Rename (presently) a single neuron or other entity (e.g. annotation)

catmaid_rename_neuron(skids = NULL, entityids = NULL, names, pid = 1,
  conn = NULL, ...)

Arguments

skids

One or more numeric skeleton ids or a character vector defining a query (see catmaid_skids or examples for the syntax).

entityids

Generic entity ids

names

New names

pid

Project id (default 1)

conn

A catmaid_connection objection returned by catmaid_login. If NULL (the default) a new connection object will be generated using the values of the catmaid.* package options as described in the help for catmaid_login.

...

Additional arguments passed to the catmaid_fetch function

Value

logical value indicating success or failure with attributes indicating the previous name.

Details

Each neuron consists of both a generic entity and an associated skeleton (skid). If you supply the skid the function will internally look up the corresponding neuron entity id. Note that you actually wished to change the name of the skeleton you could pass the skid into the entityids argument.

Examples

## Not run: ------------------------------------
# # Warning this will change data!
# n=get_neuron_name(27296)
# catmaid_rename_neuron(skids=27296, names=n)
## ---------------------------------------------