Get data.frame of connector (synapse) information from a neuron or neuronlist

In contrast to catmaid_get_connector_table this assumes that you have already read the neurons into an R structure of class neuron or neuronlist.

connectors(x, ...)

# S3 method for catmaidneuron
connectors(x, ...)

# S3 method for neuronlist
connectors(x, subset = NULL, ...)

Arguments

x

Neuron or neuronlist

...

Additional arguments passed to methods (and to nlapply in the case of connectors)

subset,

optional subset of neurons to keep (see nlapply for details)

Value

A data.frame with columns

treenode_id (NB this is the treenode id for the current skeleton)

connector_skid

prepost integer indicating whether connection is pre-(0) or post(1)-synaptic with respect to the current neuron. In other words this field will be 0 (pre) for the output synapses of this neuron.

x Spatial Location

y

z

skid For connectors.neuronlist, the skid of the skeleton from which connector information was retrieved.

See also

nlapply

Other connectors: catmaid_get_connector_table, catmaid_get_connectors_between, catmaid_get_connectors, catmaid_query_connected

Examples

## Not run: ------------------------------------
# ornsl=read.neurons.catmaid("name:ORN left", OmitFailures = T, .progress='text')
# conndf=connectors(ornsl)
# summary(connectors(ornsl))
# 
# # plot points in 3d
# library(nat)
# nopen3d()
# points3d(xyzmatrix(conndf), col=c(pre='red', post='cyan')[conndf$prepost+1])
## ---------------------------------------------