Add a 3D mesh surface to catmaid volume manager

Add a 3D mesh surface to catmaid volume manager

catmaid_add_volume(x, conn = NULL, pid = 1, ...)

Arguments

x

A mesh object such as hxsurf or a list with raw vertices and indices that will be passed to as.catmaidmesh

conn

A catmaid_connection objection returned by catmaid_login. If NULL (the default) a new connection object will be generated using the values of the catmaid.* package options as described in the help for catmaid_login.

pid

Project id (default 1)

...

Additional arguments passed to as.catmaidmesh such as title, comments etc.

Value

An invisible list with the post status response on success (the function will throw an error on failure)

See also

Examples

if (FALSE) {
library(elmr)
FAFBNP.surf=xform_brain(JFRC2NP.surf, sample=JFRC2, ref=FAFB14)
for(r in FAFBNP.surf$RegionList){
  comment="Transformed from JFRC2 space onto FAFB14"
  name=paste("v14", sep=".", r)
  catmaid_add_volume(subset(FAFBNP.surf, r), title=name, comment=comment)
}
}