Return the raw data for a CATMAID neuronal skeleton

Return the raw data for a CATMAID neuronal skeleton

catmaid_get_compact_skeleton(
  skid,
  pid = 1L,
  conn = NULL,
  connectors = TRUE,
  tags = TRUE,
  raw = FALSE,
  ...
)

Arguments

skid
single skeleton id

pid
project id (default 1)

conn
the catmaid_connection object

connectors
Whether to fetch connector information

tags
Whether to fetch tag information

raw
Whether to return completely unprocessed data (when TRUE) or to convert the nodes and connectors lists into processed data.frames (when FALSE, the default)

...
Additional arguments passed to the catmaid_fetch function.

Value

An R list object with three elements

nodes A data frame containing XYZ location, node identifiers etc for each point in the neuron.

connectors A data frame containing the position and tree node identifiers for the synaptic partners.

tags A list containing one vector for each named tag; the vectors contain node ids that are also present in the nodes element.

Details

Note that by default this fetches both skeleton and connector (synapse) information.

See also

Examples

if (FALSE) {
## ensure that you have done something like
# conn=catmaid_login()
# at least once this session to connect to the server
skel=catmaid_get_compact_skeleton(10418394)
# no connector (i.e. synapse) information
skel=catmaid_get_compact_skeleton(10418394, connectors = FALSE)

}