Get contributor statistics for neurons from CATMAID

Get contributor statistics for neurons from CATMAID

catmaid_get_contributor_stats(skids, pid = 1, conn = NULL, ...)

Arguments

skids

One or more numeric skeleton ids or a character vector defining a query (see catmaid_skids or examples for the syntax).

pid

Project id (default 1)

conn

A catmaid_connection objection returned by catmaid_login. If NULL (the default) a new connection object will be generated using the values of the catmaid.* package options as described in the help for catmaid_login.

...

Additional arguments passed to the catmaid_fetch function

Value

a list containing different statistics including construction and review times (aggregated across all the specified input neurons). There will also be 3 data.frames containing statistics for number of nodes and pre/post-synaptic connectors broken down per user.

pre_contributors number of pre-synaptic connectors contributed per user.

node_contributors number of skeleton nodes contributed per user.

post_contributors number of post-synaptic connectors contributed per user.

See also

catmaid_get_review_status

Examples

## Not run: ------------------------------------
# cs=catmaid_get_contributor_stats(skids=c(10418394,4453485))
# # fetch user list
# ul=catmaid_get_user_list()
# # merge with list of node contributors and sort in descending order of
# # contributions
# library(dplyr)
# left_join(cs$node_contributors, ul) %>%
#   select(id,n, full_name) %>%
#   arrange(desc(n))
## ---------------------------------------------