Return tree node table for a given neuron

Return tree node table for a given neuron

catmaid_get_treenode_table(skid, pid = 1, conn = NULL, raw = FALSE, ...)

Arguments

skid

Numeric skeleton id

pid

project id (default 1)

conn

the catmaid_connection object

raw

Whether to return completely unprocessed data (when TRUE) or to convert the nodes and connectors lists into processed data.frames (when FALSE, the default)

...

Additional arguments passed to the catmaid_fetch function.

Value

A data.frame with columns

id

type

tags

confidence

x

y

z

s

r

user

last_modified

reviewer (character vector with comma separated reviewer ids)

In addition two data.frames will be included as attributes: reviews, tags.

See also

catmaid_get_compact_skeleton, read.neuron.catmaid and catmaid_get_user_list to translate user ids into names.

Examples

## Not run: ------------------------------------
# # get tree node table for neuron 10418394
# tnt=catmaid_get_treenode_table(10418394)
# # show all leaf nodes
# subset(tnt, type=="L")
# # table of node types
# table(tnt$type)
# 
# # look at tags data
# str(attr(tnt, 'tags'))
# # merge with main node table to get xyz position
# tags=merge(attr(tnt, 'tags'), tnt, by='id')
# # label up a 3d neuron plot
# n=read.neuron.catmaid(10418394)
# plot3d(n, WithNodes=F)
# text3d(xyzmatrix(tags), texts = tags$tag, cex=.7)
## ---------------------------------------------