Fetch position and other information for arbitrary treenode ids

Fetch position and other information for arbitrary treenode ids

catmaid_get_treenodes_detail(tnids, pid = 1, conn = NULL, raw = FALSE,
  ...)

Arguments

tnids
One or more (integer) treenode ids

pid
project id (default 1)

conn
the catmaid_connection object

raw
Whether to return completely unprocessed data (when TRUE) or to convert the nodes and connectors lists into processed data.frames (when FALSE, the default)

...
Additional arguments passed to the catmaid_fetch function.

Value

treenode_id,

parent_id,

x,

y,

z,

confidence,

radius,

skid,

edition_time,

user_id

Details

The key feature of this function is that allows you to fetch information for arbitrary tree nodes that do not need to be from the same skeleton.

See also

Examples

# details for 3 nodes from two different skeletons
catmaid_get_treenodes_detail(c(9943214L, 25069047L, 12829015L))

#>   treenode_id parent_id       x       y     z confidence radius     skid
#> 1     9943214        NA 65439.8 20611.2 41250          5   1911  4453485
#> 2    25069047  10983327 53296.5 14791.1 42650          5     -1  4453485
#> 3    12829015  12829014 59222.2 65869.2 51150          5     -1 10418394
#>   edition_time user_id
#> 1   1496315535      18
#> 2   1525270313      71
#> 3   1523630041      18