Find neurons connected to a starting neuron

Find neurons connected to a starting neuron

catmaid_query_connected(skids, minimum_synapses = 1, boolean_op = c("OR",
  "AND"), pid = 1, raw = FALSE, conn = NULL, ...)

Arguments

skids

One or more numeric skeleton ids or a character vector defining a query (see catmaid_skids or examples for the syntax).

minimum_synapses

Must be at least this number of synapses between starter neuron and returned partners

boolean_op

Whether returned neurons can be connected to any character vector (boolean_op="OR", the default) or must be connected to all the specified neurons (boolean_op="AND").

pid

Project id (default 1)

raw

Whether to return completely unprocessed data (when TRUE) or to convert the nodes and connectors lists into processed data.frames (when FALSE, the default)

conn

A catmaid_connection objection returned by catmaid_login. If NULL (the default) a new connection object will be generated using the values of the catmaid.* package options as described in the help for catmaid_login.

...

Additional arguments passed to the catmaid_fetch function

Value

A list containing two data.frames (incoming, outgoing), each data.frame having one row for each partner neuron and the following columns

skid skeleton id of the query neuron

partner the skeleton id of the partner neuron

syn.count the number of synapses made with partner

num_nodes number of nodes in the partner neuron

See also

catmaid_skids. See catmaid_get_review_status to get information about the review status of partners (as shown in the equivalent CATMAID report).

Other connectors: catmaid_get_connector_table, catmaid_get_connectors_between, catmaid_get_connectors, connectors

Examples

## Not run: ------------------------------------
# orn13a=catmaid_query_by_name("13a ORN left")$skid
# catmaid_query_connected(orn13a)
# 
# # connected to either left OR right 13a ORNs
# orn13as=catmaid_query_by_name("13a ORN")
# catmaid_query_connected(orn13as)
# # connected to both left AND right 13a ORNs
# catmaid_query_connected(orn13as, boolean_op = 'AND')
## ---------------------------------------------